Replacing @name with ${PROJECT_NAME}
Replacing almost all occurrences of @name with ${PROJECT_NAME} in order to make changes to package name easier. The only places (excluding the project's name line at the top of the CMakeLists.txt) where @name should still be used are the source files. This will prevent changes to name of project affect the default source files defined in the CMakeLists.txt.