Fixes #130.
When generating a changelog, only tags that are ancestors of the current HEAD
should be included in the changelog. The previous behavior of
`catkin_generate_changelogs --all` was to make changelog entries for every tag
in the repository.

This commit changes the behavior of GitClient.get_tags() such that only
ancestors of the current HEAD are returned.

`git log --pretty="%d"` does not label tags in git versions less
than 1.9 (or maybe 1.8), so there is no way to distinguish between
tags and branch labels in the log output. This change uses a decorated
oneline log, which properly labels the tags. However, because the
oneline output also includes the commit hash and brief, some extra
parsing is needed to extract only the decorators.