Fixes #130.
When generating a changelog, only tags that are ancestors of the current HEAD
should be included in the changelog. The previous behavior of
`catkin_generate_changelogs --all` was to make changelog entries for every tag
in the repository.

This commit changes the behavior of GitClient.get_tags() such that only
ancestors of the current HEAD are returned.