Modifies git command to work w/ git 1.7.9
`git log --pretty="%d"` does not label tags in git versions less
than 1.9 (or maybe 1.8), so there is no way to distinguish between
tags and branch labels in the log output. This change uses a decorated
oneline log, which properly labels the tags. However, because the
oneline output also includes the commit hash and brief, some extra
parsing is needed to extract only the decorators.