Fix permission denied issue on Windows
The file created (and opened) by NamedTemporaryFile will be used in update_changelog_sections later and the second-time open() on the temp file will result in a permission denied on Windows because of how Python implementation. The issue has been discussed here: https://bugs.python.org/issue14243

And one workaround talked about in the discussion is to pass "delete=False".