Condense test matrix expression.
I don't necessarily enjoy that this splits what you have to look at
between the matrix and the includes instead of working well for one or
the other. But I think this actually comes out more intelligible and
maintainable than the big list of excludes. The goal here is:

* Test the latest versions of macOS and Ubuntu with each python version
not supported by an earlier version.
* Test each older Ubuntu version with its supported python versions.

If we wanted to be really serious about the platform-specific CI we
would install Python on the older distros from the apt repos rather than
the setup-python action so maybe this is just theatrics.