Stop using an undocumented docutils API (#308)
It looks like docutils planned to deprecate this undocumented API in
their 0.18 release, but some internal changes to avoid that deprecation
removed this field from the API.

From what I can tell, this code is trying to extract the text from the
subtitle of each subsection to parse the date and version number for
existing entries. The astext function should give us the title stripped
of formatting, which is probably what we want to parse anyway.