Mark linter test and declare cov/junit module name
Using the setup.cfg to tell coverage the name of the module we're
interested in is better than passing it on the command line every time
we invoke pytest with --cov.

Declaring the junit_suite_name will make the tests categories correctly
in our infrastructure.

Marking tests appropriately will allow us to be more selective about
what tests are run.