Add test data for a package with an xhtml description.
As mentioned in the package description itself, such formatting is, in
theory valid[1] yet it does not appear that catkin_pkg does anything
with it and in several cases it probably should.

This file is not yet wired to anything but I'm hoping to start with some
test cases regarding how xhtml ought to be handled in various scenarios
and work from there.

[1]: https://www.ros.org/reps/rep-0149.html#description