Strip spaces around newlines.
The spaces around these newlines are "valid" whitespace in a textual
rendering of the xhtml because the literal newlines around the <br>
tags would be consolidated as a single space. However, since we're
trying to produce "display quality" plaintext I think it makes sense to
try and trip spaces around newlines in order to jive with the agreed
upon "expected" result from the earlier test.