Add plaintext description field to Package. (#329)
* Add test data for a package with an xhtml description.

As mentioned in the package description itself, such formatting is, in
theory valid[1] yet it does not appear that catkin_pkg does anything
with it and in several cases it probably should.

This file is not yet wired to anything but I'm hoping to start with some
test cases regarding how xhtml ought to be handled in various scenarios
and work from there.

[1]: https://www.ros.org/reps/rep-0149.html#description

* Update test package to format 3.

* Add test for plaintext_description behavior.

When a package has an xhtml description there should be a way to get an
acceptable plaintext substitue with the xhtml tags and whitespace
stripped.

* Add stub of a method for returning the plaintext description of a package.

* Escape newlines at the start and end of multiline string.

We don't want these newlines in the expected result.

* Update description in order to exhibit additional test cases.

Also fixes a final whitespace problem with the plaintext sample.

* Call the function get_plaintext_description().

In keeping with the convention of other accessor functions like
get_build_type().

* Fix actual discrepancies between the test data and the expected result.

* Add _get_node_text function to get plaintext during xml processing.

Re-parsing the description XML made less sense than generating the
plaintext description while still processing the xml dom.
The _get_node_text recursive function walks the node tree and extracts
text from nested child nodes as opposed to _get_node_value(node,
allow_xml=False) which drops non-text children and would thus remove
emphasized text rather than render it in anyway as well as still failing
to handle inserting newlines for br tags.

This still isn't 100% matching the reference plaintext string due to
small whitespace issues and I'm trying to figure out whether there is a
generic way to deal with that.

* Fix invalid escape sequence warning by using a raw string.

* Strip spaces around newlines.

The spaces around these newlines are "valid" whitespace in a textual
rendering of the xhtml because the literal newlines around the <br>
tags would be consolidated as a single space. However, since we're
trying to produce "display quality" plaintext I think it makes sense to
try and trip spaces around newlines in order to jive with the agreed
upon "expected" result from the earlier test.

* Drop getter function since we're using an attribute.

* Move the description into a file.

I'm not pleased. But flake8 was yelling about the long-ass line and I
don't know how to noqa a docstring literal.