Add API method for getting all build_types regardless of conditions.
This is more of a spike than anything. I'm not convinced this is the
best way to go for this particular issue but it was the first thought
that came to my head.

Of possibly more concern is that this gets us past the "first hurdle"
but I'm unclear why the simple version bump doesn't crash also.