Bump the version in setup.py if ament_python is any one of the conditional build types.
There has been no previous requirement that catkin_prepare_release or
catkin_package_version be run within the context of an existing
rosdistro and I think the likelihood that we would end up being run in
the "wrong" context is higher than the benefit from being run in
context.

For a single-source package using conditional build types it isn't even
clear what the "correct" context is? Maintainers would need to be aware
that running in a context that evaluates the build type to ament_python
(currently) triggers a superset of the behaviors for other build types.

This problem will only get worse as different build types require
different, possibly (albeit unlikely) conflicting transformations in
order to produce a release / version bump.

I think it's easier on the maintainer to do the right thing for all
build types rather than picking one and try to do the right thing for
it.