Declare dependency on setuptools in stdeb. (#360)
setuptools was previously all but guaranteed on python systems even if
it was never explicitly required. It's possible to install catkin_pkg on
a noble system which lacks setuptools.

I believe we're using it for the entrypoints behavior and there may be a
stdlib or other alternative dependency we can use in the future.