Update rather than overwrite _ansi module variable.
The flake8 global complaint () has previously caught cases where the
global declaration was unecessary. In this case, a test started failing
after the change and I'm struggling to understand how the previous
behavior was working properly since as far as I can tell it does
overwrite the global rather than update it.